scrapper/requirements.txt
attrs==23.1.0 ; python_version >= "3.9" and python_version < "4.0"
beautifulsoup4==4.12.2 ; python_version >= "3.9" and python_version < "4.0"
bs4==0.0.1 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.15.1 ; os_name == "nt" and implementation_name != "pypy" and python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
exceptiongroup==1.1.3 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
html-to-json==2.0.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.25.2 ; python_version < "4.0" and python_version >= "3.9"
outcome==1.2.0 ; python_version >= "3.9" and python_version < "4.0"
pandas==2.0.3 ; python_version >= "3.9" and python_version < "4.0"
pillow==10.0.0 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; os_name == "nt" and implementation_name != "pypy" and python_version >= "3.9" and python_version < "4.0"
pysocks==1.7.1 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
selenium==4.11.2 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.9" and python_version < "4.0"
soupsieve==2.4.1 ; python_version >= "3.9" and python_version < "4.0"
trio-websocket==0.10.3 ; python_version >= "3.9" and python_version < "4.0"
trio==0.22.2 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.0.4 ; python_version >= "3.9" and python_version < "4.0"
urllib3[socks]==2.0.4 ; python_version >= "3.9" and python_version < "4.0"
wsproto==1.2.0 ; python_version >= "3.9" and python_version < "4.0"
